- Fix psworld. - Some warning fixes.

+++ trunk/src/server/adminmanager.cpp
@@ -6433,7 +6433,7 @@ void AdminManager::HandleListWarnings(psAdminCmdMessage& msg, AdminCmdData& data
         if (rs.IsValid())
         {
             csString newLine;
-            size_t i = 0;
+            unsigned long i = 0;
             for (i = 0 ; i < rs.Count() ; i++)
             {
                 newLine.Format("%s - %s - %s", rs[i]["warningGM"], rs[i]["timeOfWarn"], rs[i]["warnMessage"]);
